[MIG] web_widget_mpld3_chart: Migration to 17.0
