+++ web_widget_mpld3_chart/__manifest__.py
@@ -6,11 +6,11 @@
     "category": "Hidden",
     "summary": "This widget allows to display charts using MPLD3 library.",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
-    "version": "16.0.1.0.0",
+    "version": "17.0.1.0.0",
     "website": "https://github.com/OCA/web",
     "depends": ["web"],
     "data": [],
-    "external_dependencies": {"python": ["mpld3==0.5.9", "beautifulsoup4"]},
+    "external_dependencies": {"python": ["mpld3==0.5.10", "beautifulsoup4"]},
     "auto_install": False,
     "development_status": "Beta",
     "maintainers": ["JordiBForgeFlow", "ChrisOForgeFlow"],
