+++ web_widget_mpld3_chart/static/src/xml/web_widget_mpld3_chart.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates xml:space="preserve">
-    <div t-ref="widget" t-name="web_widget_mpld3_chart.Mpld3ChartJsonWidget" owl="1">
-        <t t-if="props.value.div" t-out="markup(props.value.div)" />
+    <div t-ref="widget" t-name="web_widget_mpld3_chart.Mpld3ChartJsonWidget">
+        <t
+            t-if="this.props.record.data[this.props.name].div"
+            t-out="markup(this.props.record.data[this.props.name].div)"
+        />
     </div>
 </templates>
